// src/app/models/employee.model.ts
export interface Employee {
    id: any;
    employeeId: number;
    userName: string;
    firstName: string;
    lastName: string;
    email: string;
    phoneNumber: string;
    dateOfBirth: string;
    hireDate: string;
    departmentId: number;
    departmentName: string;
    position: string;
    city: string;
    jobTitle: string;
    profileImagePath: string | null;
    role: string;
  }